--- src/@types/arweaveFileObject.ts
@@ -1,4 +1,0 @@
-export interface ArweaveFileObject {
-  type: string
-  transactionId: string
-}
