--- src/@types/ipfsFileObject.ts
@@ -1,4 +1,0 @@
-export interface IpfsFileObject {
-  type: string
-  hash: string
-}
